手持ちポケモン機能&最低限のComponent管理s
