+++ PokemonMaster/components/FilteredArray.vue
@@ -1,10 +1,12 @@
-/* eslint-disable no-console */
 <template>
   <v-container>
-    <p>お気に入りポケモンのページ</p>
+    <v-row>
+      <p v-if="filteredPokemons.length == 0">検索結果は0件でした。</p>
+      <p v-else>検索結果は{{ filteredPokemons.length }}件ありました。</p>
+    </v-row>
     <v-row>
       <v-col
-        v-for="pokemon in pokemons"
+        v-for="pokemon in filteredPokemons"
         :key="pokemon.id"
         xs="6"
         sm="3"
@@ -15,7 +17,7 @@
           min-width="140px"
           max-width="150px"
           :to="{ name: 'pokemons-id', params: { id: pokemon.id } }"
-          ><v-img :src="pokemon.srcNormal"></v-img>
+          ><v-img :src="pokemon.src.normal"></v-img>
           <v-card-title>{{ pokemon.name }}</v-card-title>
           <v-card-subtitle v-if="pokemon.type.length == 1">{{
             pokemon.type[0]
@@ -39,38 +41,38 @@
 </template>
 
 <script>
-// import axios from 'axios'
+import axios from 'axios'
 import firebase from '~/plugins/firebase'
 export default {
   data() {
     return {
+      type: true,
+      valid: true,
       keyword: '',
-      favorites: [],
       pokemons: [],
     }
   },
+  computed: {
+    filteredPokemons() {
+      const pokemons = []
+      for (const i in this.pokemons) {
+        const pokemon = this.pokemons[i]
+        if (pokemon.name.includes(this.$route.query.name)) {
+          pokemons.push(pokemon)
+        }
+      }
+      return pokemons
+    },
+  },
   mounted() {
-    const that = this
-    const db = firebase.firestore()
-    const dbFavorites = db.collection('favorites').orderBy('createdAt', 'desc')
-    dbFavorites.get().then((snapshot) => {
-      snapshot.forEach((doc) => {
-        const favorites = doc.data()
-        that.pokemons = [
-          ...that.pokemons,
-          {
-            id: favorites.id,
-            name: favorites.name,
-            type: favorites.type,
-            srcIcon: favorites.srcIcon,
-            srcNormal: favorites.srcNormal,
-            srcSmall: favorites.srcSmall,
-            createdAt: favorites.createdAt,
-            updatedAt: favorites.updatedAt,
-          },
-        ]
+    axios
+      .get('../pokemon.json')
+      .then((response) => {
+        this.pokemons = response.data
       })
-    })
+      .catch((err) => {
+        alert('エラー:' + err)
+      })
   },
   methods: {
     toPokemon() {
@@ -92,9 +94,9 @@ export default {
           id: pokemon.id,
           name: pokemon.name,
           type: pokemon.type,
-          srcIcon: pokemon.srcIcon,
-          srcNormal: pokemon.srcNormal,
-          srcSmall: pokemon.srcSmall,
+          srcIcon: pokemon.src.icon,
+          srcNormal: pokemon.src.normal,
+          srcSmall: pokemon.src.small,
           createdAt: timestamp,
           updatedAt: timestamp,
         })
